Reorganize the applicative layer code.
Apps are now in different files, and register via decorators to a
central registry. The MinitelHandler relies on the registry to handle
dispatch.

It is also now possible to create "forking" apps, i.e. apps that shell
out to another binary. The binary is invoked with TERM=minitel so this
only works for apps that properly handle termcaps (unlike most of the
world which hardcodes ANSII escape sequences). An example app is
provided using Nethack, which is really cool to play on a minitel.