Add xtel module and package.
The package was tricky as this is 2001 software, but it compiles fine
with minimal changes. Most changes are related to xteld's opinionated
handling of directories (make install wanted to write directly to
/etc/ with no support for --prefix). Instead, we add the
xtel.{lignes,services} to $out/usr/share as "examples".

The module runs xteld under xinetd. By default the example configs are
symlinked to /etc/ for consumption by xteld, however users would
likely redefeine at least the xtel.lignes file to point to their own
modem.

Overlays are provided to inherit xtel, minitel-server, or both.