Bundle the web, websocket and tty services in the same main.
Reflect these changes in the NixOS module.

This means that the main can now start a tty service, a webserver
serving the emulator (and eventually the flask-admin app), and the
websocket required to serve the minitel app on the emulator. This
makes it much easier to iterate on the app contents.